Extract module test.on_yubikey.cli_piv.test_pin_puk
